pyproject.toml
[project]
name = "airflow-test2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.2.0",
]
